+++ modules/datastore/tests/src/Functional/DictionaryEnforcerTest.php
@@ -6,10 +6,12 @@ use Drupal\Core\File\FileSystemInterface;
 
 use Drupal\datastore\Controller\ImportController;
 use Drupal\metastore\DataDictionary\DataDictionaryDiscovery;
+use Drupal\Tests\BrowserTestBase;
 use Drupal\Tests\common\Traits\CleanUp;
 use Drupal\Tests\common\Traits\GetDataTrait;
 use Drupal\Tests\metastore\Unit\MetastoreServiceTest;
 
+use RootedData\RootedJsonData;
 use Symfony\Component\HttpFoundation\Request;
 use weitzman\DrupalTestTraits\ExistingSiteBase;
 
@@ -18,10 +20,19 @@ use weitzman\DrupalTestTraits\ExistingSiteBase;
  *
  * @package Drupal\Tests\datastore\Functional
  * @group datastore
+ * @group functional
+ * @group btb
  */
-class DictionaryEnforcerTest extends ExistingSiteBase {
+class DictionaryEnforcerBTBTest extends BrowserTestBase {
 
   use GetDataTrait, CleanUp;
+
+  protected $defaultTheme = 'stark';
+
+  protected static $modules = [
+    'datastore',
+    'node',
+  ];
 
   /**
    * Uploaded resource file destination.
@@ -84,7 +95,7 @@ class DictionaryEnforcerTest extends ExistingSiteBase {
    *
    * @var \Drupal\datastore\Controller\ImportController
    */
-  protected $webServiceApi;
+  protected $importController;
 
   /**
    * External URL for the fixture CSV file.
@@ -100,35 +111,29 @@ class DictionaryEnforcerTest extends ExistingSiteBase {
     parent::setUp();
 
     // Initialize services.
-    $this->cron = \Drupal::service('cron');
-    $this->metastore = \Drupal::service('dkan.metastore.service');
-    $this->uuid = \Drupal::service('uuid');
+    $this->cron = $this->container->get('cron');
+    $this->metastore = $this->container->get('dkan.metastore.service');
+    $this->uuid = $this->container->get('uuid');
     $this->validMetadataFactory = MetastoreServiceTest::getValidMetadataFactory($this);
-    $this->webServiceApi = ImportController::create(\Drupal::getContainer());
-    $this->datasetStorage = \Drupal::service('dkan.metastore.storage')
+    $this->importController = ImportController::create(\Drupal::getContainer());
+    $this->datasetStorage = $this->container->get('dkan.metastore.storage')
       ->getInstance('dataset');
     // Copy resource file to uploads directory.
     /** @var \Drupal\Core\File\FileSystemInterface $file_system */
-    $file_system = \Drupal::service('file_system');
+    $file_system = $this->container->get('file_system');
     $upload_path = $file_system->realpath(self::UPLOAD_LOCATION);
     $file_system->prepareDirectory($upload_path, FileSystemInterface::CREATE_DIRECTORY);
     $file_system->copy(self::TEST_DATA_PATH . self::RESOURCE_FILE, $upload_path, FileSystemInterface::EXISTS_REPLACE);
-    // Create https resource URL.
-    $this->resourceUrl = \Drupal::service('file_url_generator')
-      ->generateAbsoluteString(self::UPLOAD_LOCATION . self::RESOURCE_FILE);
-  }
-
-  public function tearDown(): void {
-    parent::tearDown();
-    $this->removeAllMappedFiles();
+    // Create resource URL.
+    $this->resourceUrl = $this->container->get('stream_wrapper_manager')
+      ->getViaUri(self::UPLOAD_LOCATION . self::RESOURCE_FILE)
+      ->getExternalUrl();
   }
 
   /**
    * Test dictionary enforcement.
    */
   public function testDictionaryEnforcement(): void {
-    // Clear out the cron queue ahead of time.
-    $this->cron->run();
     // Build data-dictionary.
     $dict_id = $this->uuid->generate();
     $fields = [
@@ -182,47 +187,47 @@ class DictionaryEnforcerTest extends ExistingSiteBase {
       $dict_id,
       $this->metastore->post('data-dictionary', $data_dict)
     );
-    $this->metastore->publish('data-dictionary', $dict_id);
-    // Check the round-trip of the dictionary.
-    $rooted_dictionary = $this->metastore->get('data-dictionary', $dict_id, FALSE);
-    $this->assertEquals(
-      $dict_id,
-      json_decode($rooted_dictionary->pretty())->identifier
-    );
+    // Publish should return FALSE, because the node was already published.
+    $this->assertFalse($this->metastore->publish('data-dictionary', $dict_id));
 
     // Set global data-dictinary in metastore config.
-    $metastore_config = \Drupal::configFactory()
-      ->getEditable('metastore.settings');
-    $metastore_config->set('data_dictionary_mode', DataDictionaryDiscovery::MODE_SITEWIDE);
-    $metastore_config->set('data_dictionary_sitewide', $dict_id);
-    $metastore_config->save();
+    $metastore_config = $this->config('metastore.settings');
+    $metastore_config->set('data_dictionary_mode', DataDictionaryDiscovery::MODE_SITEWIDE)
+      ->set('data_dictionary_sitewide', $dict_id)
+      ->save();
 
     // Build dataset.
     $dataset_id = $this->uuid->generate();
-    $dataset = $this->validMetadataFactory->get($this->getDataset($dataset_id, 'Test ' . $dataset_id, [$this->resourceUrl], TRUE), 'dataset');
+    $this->assertInstanceOf(
+      RootedJsonData::class,
+      $dataset = $this->validMetadataFactory->get(
+        $this->getDataset($dataset_id, 'Test ' . $dataset_id, [$this->resourceUrl], TRUE),
+        'dataset'
+      )
+    );
     // Create dataset.
     $this->assertEquals($dataset_id, $this->metastore->post('dataset', $dataset));
     // Publish should return FALSE, because the node was already published.
-    $this->assertEquals(FALSE, $this->metastore->publish('dataset', $dataset_id));
+    $this->assertFalse($this->metastore->publish('dataset', $dataset_id));
 
-    // Run queues to import the dataset and apply the dictionary.
+    // Run queue items to perform the import.
     $this->runQueues(['datastore_import', 'post_import']);
 
     // Retrieve dataset distribution ID.
-    /** @var \RootedData\RootedJsonData $dataset */
-    $dataset = $this->metastore->get('dataset', $dataset_id);
-//    $this->assertEquals('foo', $dataset->pretty());
-    $dist_uuid = $dataset->{'$["%Ref:distribution"][0].identifier'} ?? NULL;
+    $this->assertInstanceOf(
+      RootedJsonData::class,
+      $dataset = $this->metastore->get('dataset', $dataset_id)
+    );
+    $this->assertNotEmpty(
+      $dist_id = $dataset->{'$["%Ref:distribution"][0].identifier'} ?? NULL
+    );
     // Retrieve schema for dataset resource.
-    $response = $this->webServiceApi->summary(
-      $dist_uuid,
-      Request::create('http://blah//api/1/datastore/imports/' . $dist_uuid)
+    $response = $this->importController->summary(
+      $dist_id,
+      Request::create('http://blah/api')
     );
     $this->assertEquals(200, $response->getStatusCode(), $response->getContent());
     $result = json_decode($response->getContent(), TRUE);
-
-    // Clean up after ourselves, before performing the assertion.
-    $this->metastore->delete('dataset', $dataset_id);
 
     // Validate schema.
     $this->assertEquals([
@@ -279,13 +284,17 @@ class DictionaryEnforcerTest extends ExistingSiteBase {
     ], $result);
   }
 
+  /**
+   * Process queues in a predictable order.
+   */
   private function runQueues(array $relevantQueues = []) {
     /** @var \Drupal\Core\Queue\QueueWorkerManager $queueWorkerManager */
     $queueWorkerManager = \Drupal::service('plugin.manager.queue_worker');
-    $queueService = \Drupal::service('queue');
+    /** @var \Drupal\Core\Queue\QueueFactory $queueFactory */
+    $queueFactory = $this->container->get('queue');
     foreach ($relevantQueues as $queueName) {
       $worker = $queueWorkerManager->createInstance($queueName);
-      $queue = $queueService->get($queueName);
+      $queue = $queueFactory->get($queueName);
       while ($item = $queue->claimItem()) {
         $worker->processItem($item->data);
         $queue->deleteItem($item);
